+++ App.js
@@ -3,7 +3,7 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createMaterialBottomTabNavigator } from "@react-navigation/material-bottom-tabs";
 import MaterialCommunityIcons from "react-native-vector-icons/MaterialCommunityIcons";
 import Home from "./src/features/home/screens/Home";
-import Course from "./src/features/course/screens/Course";
+import Courses from "./src/features/course/screens/Courses";
 import Schedule from "./src/features/schedule/screens/Schedule";
 
 const Tab = createMaterialBottomTabNavigator();
@@ -24,7 +24,7 @@ export default function App() {
         />
         <Tab.Screen
           name="Course"
-          component={Course}
+          component={Courses}
           options={{
             tabBarLabel: "Course",
             tabBarIcon: ({ color }) => (
